LayoutTests/platform/chromium/fast/events/touch/gesture/gesture-scroll-expected.txt

This tests scroll gesture events. Square is (mostly) green on pass

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


first gesture
PASS movingdiv.scrollTop is 0
PASS movingdiv.scrollLeft is 90
second gesture
PASS movingdiv.scrollTop is 95
PASS movingdiv.scrollLeft is 90
PASS successfullyParsed is true

TEST COMPLETE
